ai/ag_math_solver/requirements.txt
streamlit==1.32.0
pyautogen==0.1.14
python-dotenv==1.0.0
openai==0.28.1
tiktoken==0.5.1
tavily-python==0.2.8
sqlite3-api==1.0.1
matplotlib==3.8.2
sympy==1.12.0
numpy==1.26.3
pandas==2.1.4
matplotlib==3.8.2
